Rename sir.cpp to model.cpp Rename SIR folder to Model
Rename sir to sir-model in cmake

mods
