+++ src/CMakeLists.txt
@@ -1,6 +1,6 @@
 ADD_EXECUTABLE(sir-sym main.cpp)
 
-ADD_SUBDIRECTORY(SIR)
+ADD_SUBDIRECTORY(Model)
 
-TARGET_LINK_LIBRARIES(sir-sym sir)
+TARGET_LINK_LIBRARIES(sir-sym sir-model)
 TARGET_LINK_LIBRARIES(sir-sym sfml-graphics)
