+++ src/Model/CMakeLists.txt
@@ -1,0 +1,1 @@
+ADD_LIBRARY(sir-model STATIC model.cpp)
