--- src/SIR/CMakeLists.txt
@@ -1,1 +1,0 @@
-ADD_LIBRARY(sir STATIC sir.cpp)
